Temporarily say goodbye to clippy
Clippy is a super useful tool. However, being a compiler plugin, it's bound
to be frequently broken as changes are made to the Rust parser.

Contributing fixes to Clippy is neat, but trying to ensure that Flowgger
always compiles with the current nightly instead of a specific snapshot is
also a good property to have.

We will eventually switch to cargo-clippy, which is a less intrusive way to
keep using clippy.

Bump serde's version by the way.