New input stream type: Redis queues
This uses the Redis reliable queue pattern, with a secondary key
each entry gets moved to before being processed.

Entries are currently being processed one by one, in a single task
(which is trivial to make concurrent, if needed). Batch processing
may be implemented later on if it makes a significant change in
performance on actual workloads.