Add tls_co, an optional, coroutines-based TLS input processor
Call that a thread pool if Java is your thing.

This requires a fork of coio, which comes with dependencies that are bound to
break regularly. It's slow, it's fragile, so much could be factored with the
regular TLS implementation but hasn't yet, and we're now trapped in a
dependency hell.

I still have great hopes that this horror won't be much than a temporary
kludge.

If you know why that had to be added, rejoice. If you don't, ignore.