Import a Cap'n Proto schema.
The ability to send Flowgger's output to another Flowgger server can be
useful.

However, using a slow encoding such as GELF between to relay messages
between both instances would be dumb.

We still need to keep track of the values types, though.

Cap'n Proto is a perfect serialization format for that.
So, this is a baby step towards supporting Cap'n Proto.