Introduce yet another node to the processing chain: the Merger
A Merger takes a record and, with relevant output types (i.e. not Kafka),
wraps the record with whatever it takes to split it.

In the configuration file, this adds a new "framing" option to the
[output] section, and obviously is the output counterpart of the [input]
framing parameter.

Currently supported values:
- framing = "noop" - Do not wrap record. Usually what you want with binary
formats such as msgpack or capnp.
- framing = "line" - Appends \n to encoded records.
- framing = "nul" - Appends \0 to encoded records.
- framing = "syslen" - Prepends the size to messages, as documented in RFC6587.

It uses sensible defaults according to the output type and format.

This extra step is required to send GELF data over TCP, where we need nul
framing, whereas the exact same GELF data should not be framed when sent to
Kafka. And when sent to the debug output, we probably want \n instead.