Statically compile LibreSSL.
Yup, the resulting binary is now pretty large, but at the end of the
day, we get a single binary that we can easily deploy anywhere.

Like, on machines still running a dead old OpenSSL, and on distros that
like to use funny versioning systems that are incompatible with everything
else.

Also, don't even try tweaking sysctls in the Docker container.
It just doesn't work, and TCP keepalive wasn't the cause of the problem
we were initially investigating.